requirements.txt
python-dotenv
streamlit
langchain_groq
langchain_community
langchain
langchain_core
pypdf
langchain-nvidia-ai-endpoints
pinecone
pinecone_text